[Behat] Extract ProductAttributesTrait from product pages
