+++ src/Sylius/Behat/Page/Admin/Product/Common/ProductAttributesTrait.php
@@ -1,0 +1,175 @@
+<?php
+
+/*
+ * This file is part of the Sylius package.
+ *
+ * (c) Sylius Sp. z o.o.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Sylius\Behat\Page\Admin\Product\Common;
+
+use Sylius\Behat\Service\DriverHelper;
+
+trait ProductAttributesTrait
+{
+    public function getDefinedProductAttributesElements(): array
+    {
+        return [
+            'product_attribute_delete_button' => '[data-test-product-attribute-delete-button="%attributeName%"]',
+            'product_attribute_input' => 'input[name="product_attributes"]',
+            'product_attribute_tab' => '[data-test-product-attribute-tab="%name%"]',
+            'attribute_value' => '[data-test-attribute-value][data-test-locale-code="%localeCode%"][data-test-attribute-name="%attributeName%"]',
+        ];
+    }
+
+    public function getAttributeValidationErrors(string $attributeName, string $localeCode): string
+    {
+        $this->clickTabIfItsNotActive('attributes');
+
+        $validationError = $this->getElement('attribute')->find('css', '.sylius-validation-error');
+
+        return $validationError->getText();
+    }
+
+    public function removeAttribute(string $attributeName, string $localeCode): void
+    {
+        $this->changeTab('attributes');
+
+        $this->getElement('product_attribute_delete_button', ['%attributeName%' => $attributeName])->press();
+
+        $this->waitForFormUpdate();
+    }
+
+    public function getAttributeSelectText(string $attribute, string $localeCode): string
+    {
+        $this->clickTabIfItsNotActive('attributes');
+
+        return $this->getElement('attribute_select', ['%attributeName%' => $attribute, '%localeCode%' => $localeCode])->getText();
+    }
+
+    public function getNonTranslatableAttributeValue(string $attribute): string
+    {
+        $this->clickTabIfItsNotActive('attributes');
+
+        return $this->getElement('non_translatable_attribute', ['%attributeName%' => $attribute])->getValue();
+    }
+
+    public function hasAttribute(string $attributeName): bool
+    {
+        return null !== $this->getDocument()->find('css', sprintf('.attribute .label:contains("%s")', $attributeName));
+    }
+
+    public function hasNonTranslatableAttributeWithValue(string $attributeName, string $value): bool
+    {
+        $attribute = $this->getDocument()->find('css', sprintf('.attribute .attribute-label:contains("%s")', $attributeName));
+
+        return
+            $attribute->getParent()->getParent()->find('css', '.attribute-input input')->getValue() === $value &&
+            $attribute->find('css', '.globe.icon') !== null
+            ;
+    }
+
+    public function addNonTranslatableAttribute(string $attributeName, string $value): void
+    {
+        $this->clickTabIfItsNotActive('attributes');
+
+        $attributeOption = $this->getElement('attributes_choice')->find('css', sprintf('option:contains("%s")', $attributeName));
+        $this->selectElementFromAttributesDropdown($attributeOption->getAttribute('value'));
+
+        $this->getDocument()->pressButton('Add attributes');
+        $this->waitForFormElement();
+
+        $this->getElement('non_translatable_attribute_value', ['%attributeName%' => $attributeName])->setValue($value);
+    }
+
+    public function addAttribute(string $attributeName): void
+    {
+        $this->changeTab('attributes');
+        $this->selectAttributeToBeAdded($attributeName);
+        $this->clickButton('Add');
+
+        $this->waitForFormUpdate();
+    }
+
+    public function updateAttribute(string $attributeName, string $value, string $localeCode): void
+    {
+        $this->changeTab('attributes');
+        $this->changeAttributeTab($attributeName);
+
+        $attributeValue = $this->getElement('attribute_value', ['%attributeName%' => $attributeName, '%localeCode%' => $localeCode]);
+
+        match ($attributeValue->getTagName()) {
+            'input' => $attributeValue->setValue($value),
+            'select' => $attributeValue->selectOption($value),
+            default => throw new \InvalidArgumentException('Unsupported attribute value type'),
+        };
+
+        $attributeValue->blur();
+        $this->waitForFormUpdate();
+    }
+
+    public function getAttributeValue(string $attribute, string $localeCode): string
+    {
+        $this->changeTab('attributes');
+        $this->changeAttributeTab($attribute);
+
+        $attributeValue = $this->getElement('attribute_value', ['%attributeName%' => $attribute, '%localeCode%' => $localeCode]);
+
+        return match ($attributeValue->getTagName()) {
+            'input' => $attributeValue->getValue(),
+            'select' => $attributeValue->getText(),
+            default => throw new \InvalidArgumentException('Unsupported attribute value type'),
+        };
+    }
+
+    public function addSelectedAttributes(): void
+    {
+        $this->changeTab('attributes');
+        $this->clickButton('Add');
+        $this->waitForFormUpdate();
+    }
+
+    public function getNumberOfAttributes(): int
+    {
+        return count($this->getDocument()->findAll('css', '[data-test-product-attribute-tab]'));
+    }
+
+    private function selectAttributeToBeAdded(string $attributeName): void
+    {
+        $driver = $this->getDriver();
+        $this->autocompleteHelper->selectByName(
+            $driver,
+            $this->getElement('product_attribute_input')->getXpath(),
+            $attributeName,
+        );
+    }
+
+    private function clickTabIfItsNotActive(string $tabName): void
+    {
+        $attributesTab = $this->getElement('tab', ['%name%' => $tabName]);
+        if (!$attributesTab->hasClass('active')) {
+            $attributesTab->click();
+        }
+    }
+
+    private function changeAttributeTab(string $attributeName): void
+    {
+        if (DriverHelper::isNotJavascript($this->getDriver())) {
+            return;
+        }
+
+        $this->getElement('product_attribute_tab', ['%name%' => $attributeName])->click();
+    }
+
+    private function clickButton(string $locator): void
+    {
+        if (DriverHelper::isJavascript($this->getDriver())) {
+            $this->getDocument()->pressButton($locator);
+        }
+    }
+}
